app/requirements.txt
-i https://pypi.tuna.tsinghua.edu.cn/simple
asn1crypto==0.24.0
blinker==1.4
bootstrap-flask==1.0.10
cffi==1.12.3
click==7.0
cryptography==2.6.1
flask-avatars==0.2.2
flask-dropzone==1.5.3
flask-login==0.4.1
flask-mail==0.9.1
flask-moment==0.7.0
flask-sqlalchemy==2.3.2
flask-whooshee==0.6.0
flask-wtf==0.14.2
flask==1.0.2
itsdangerous==1.1.0
jinja2==2.10.1
markupsafe==1.1.1
pillow==6.0.0
pycparser==2.19
pymysql==0.9.3
python-dotenv==0.10.1
six==1.12.0
sqlalchemy==1.3.3
werkzeug==0.15.2
whoosh==2.7.4
wtforms==2.2.1
